src/utils/config.py
"""
Configuration management for the Sidekick AI application.
Handles environment variables and application settings.
"""

import os
from typing import Optional
from dotenv import load_dotenv

# Load environment variables
load_dotenv(override=True)


class Config:
    """Application configuration class."""
    
    # OpenAI Configuration
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY", "")
    
    # Pushover Configuration (Optional)
    PUSHOVER_TOKEN: Optional[str] = os.getenv("PUSHOVER_TOKEN")
    PUSHOVER_USER: Optional[str] = os.getenv("PUSHOVER_USER")
    
    # Google Search API (Optional)
    SERPER_API_KEY: Optional[str] = os.getenv("SERPER_API_KEY")
    
    # Application Configuration
    APP_ENV: str = os.getenv("APP_ENV", "development")
    APP_DEBUG: bool = os.getenv("APP_DEBUG", "true").lower() == "true"
    APP_HOST: str = os.getenv("APP_HOST", "0.0.0.0")
    APP_PORT: int = int(os.getenv("APP_PORT", "7860"))
    
    # Logging Configuration
    LOG_LEVEL: str = os.getenv("LOG_LEVEL", "INFO")
    LOG_FILE: str = os.getenv("LOG_FILE", "logs/sidekick.log")
    
    @classmethod
    def validate(cls) -> bool:
        """Validate required configuration."""
        if not cls.OPENAI_API_KEY:
            raise ValueError("OPENAI_API_KEY is required")
        return True
    
    @classmethod
    def is_production(cls) -> bool:
        """Check if running in production mode."""
        return cls.APP_ENV.lower() == "production"
    
    @classmethod
    def is_development(cls) -> bool:
        """Check if running in development mode."""
        return cls.APP_ENV.lower() == "development"


# Global configuration instance
config = Config()
